requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# TRELLIS dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0
xformers>=0.0.20
flash-attn>=2.3.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
imageio>=2.31.0
rembg>=2.0.50

# 3D processing
trimesh>=4.0.0
pygltflib>=1.16.0
plyfile>=0.7.4

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Redis
redis>=5.0.0
celery>=5.3.0

# Payment processing
stripe>=7.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# File storage
boto3>=1.28.0
minio>=7.1.0

# HTTP client
httpx>=0.25.0
aiofiles>=23.2.0

# Monitoring
prometheus-client>=0.19.0
structlog>=23.2.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
isort>=5.12.0
mypy>=1.6.0

# Telegram Bot
python-telegram-bot>=20.7

# Web interface
gradio>=4.0.0
